+++ cate/util/time.py
@@ -30,6 +30,7 @@ _RE_TO_DATETIME_FORMATS = patterns = [(re.compile(14 * '\\d'), '%Y%m%d%H%M%S'),
                                       (re.compile(6 * '\\d'), '%Y%m'),
                                       (re.compile(4 * '\\d'), '%Y')]
 
+
 def find_datetime_format(filename: str) -> Tuple[Optional[str], int, int]:
     for regex, time_format in _RE_TO_DATETIME_FORMATS:
         searcher = regex.search(filename)
@@ -37,6 +38,7 @@ def find_datetime_format(filename: str) -> Tuple[Optional[str], int, int]:
             p1, p2 = searcher.span()
             return time_format, p1, p2
     return None, -1, -1
+
 
 def get_timestamp_from_string(string: str) -> pd.Timestamp:
     time_format, p1, p2 = find_datetime_format(string)
@@ -46,6 +48,7 @@ def get_timestamp_from_string(string: str) -> pd.Timestamp:
         return pd.to_datetime(string[p1:p2], format=time_format)
     except ValueError:
         return None
+
 
 def get_timestamps_from_string(string: str) -> (pd.Timestamp, pd.Timestamp):
     first_time = None